refactor(frame): pin으로 canroll 역할 수정
