+++ src/main/java/bowling/domain/pin/Pins.java
@@ -1,5 +1,6 @@
 package bowling.domain.pin;
 
+import bowling.domain.frame.FinalFrame;
 import bowling.domain.frame.Frame;
 import bowling.domain.frame.NormalFrame;
 
@@ -7,13 +8,12 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Pins {
-    private static final int PINS_LIMIT = 10;
-
-    private static final int FIRST_FRAME = 1;
-    private static final int FINAL_FRAME = 10;
     private static final String PIN_MAX_ERROR = "핀의 합계가 10개보다 클 수 없습니다.";
-    private static final int CAN_ROLL_LIMIT = 2;
+    private static final String ROLL_COUNT_ERRORS = "일반 게임에서는 두번만 던질 수 있습니다.";
+    private static final int NORMAL_FRAME_CAN_ROLL = 2;
+    private static final int FINAL_FRAME_CAN_ROLL = 3;
     private static final int FIRST_ROLL = 1;
+    private static final int PINS_LIMIT = 10;
 
     private List<Pin> pins;
 
@@ -32,15 +32,8 @@ public class Pins {
                 .sum();
     }
 
-
-
     public int getLeftPin() {
         return PINS_LIMIT - getTotalPins();
-    }
-
-
-    public boolean isRolledTwice() {
-        return pins.size() >= CAN_ROLL_LIMIT;
     }
 
     public boolean isAlreadyStrike() {
@@ -51,15 +44,38 @@ public class Pins {
         return this.getTotalPins() + pin > PINS_LIMIT;
     }
 
-    public boolean isPinReady() { return !isAlreadyStrike() && !isRolledTwice();}
+    public boolean isPinReady(Frame frame) {
+        if (frame instanceof NormalFrame) {
+            return !isAlreadyStrike() && !isRolledTwice();
+        }
+        if (!isRolledTwice() || getTotalPins() >= PINS_LIMIT) {
+            return pins.size() < FINAL_FRAME_CAN_ROLL;
+        }
+        return false;
+    }
 
     public void addPins(Frame frame, int pin) {
-
         if (frame instanceof NormalFrame && isPinTotalOverTen(pin)) {
             throw new IllegalArgumentException(PIN_MAX_ERROR);
         }
 
+        if (frame instanceof FinalFrame && isPinOverflowRolledOnce(pin)) {
+            throw new IllegalArgumentException(ROLL_COUNT_ERRORS);
+        }
+
         this.pins.add(new Pin(pin));
+    }
+
+    private boolean isPinOverflowRolledOnce(int pin) {
+        return isRolledOnce() && (isPinUnderTen()) && (getTotalPins() + pin) > PINS_LIMIT;
+    }
+
+    private boolean isRolledTwice() {
+        return pins.size() == NORMAL_FRAME_CAN_ROLL;
+    }
+
+    private boolean isRolledOnce() {
+        return pins.size() < NORMAL_FRAME_CAN_ROLL;
     }
 
     public List<Pin> getPins() {
@@ -74,6 +90,9 @@ public class Pins {
         return this.pins.size();
     }
 
+    private boolean isPinUnderTen() {
+        return getTotalPins() < PINS_LIMIT;
+    }
 
 
 }
